Q2/requirements.txt
Django==3.0.6
django-filter==2.2.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
mysqlclient==1.4.6
PyMySQL==0.9.3
